msp430: move clock spec units from KHZ to HZ to avoid confusion.
Motivation: TinyOS timing is defined to be binary time and binary frequencies.
Primarily because the master clock source is a 32768 Hz (power of 2) crystal.
But there are many places where sometime MHz means power of ten and other time
when it is a binary MHz (power of 2).   Remove that ambiguity.  binary MHz -> MiHz
and binary KHz -> KiHz.

Also change the MSP DCO calculation to Hz from binary KHz (KiHz) to further avoid
any ambiguity.