msp430: make configurations be in ROM (const)
Motivation:  Without the const modifier, the configuration structures will
be in RAM and the actual contents will be in ROM.   Double the space plus
the cost of copying down the data.  Further these structures being in RAM
are subject to corruption.  Seems silly.   Force and use the structures
out of ROM.

Note eyesIFX modifies its structure.  So we leave that in RAM and cast it.