msp430: rename usci ports, ie. mv UART0 -> UARTA0, etc.
Motivation:  The original MSP430 port defined i/o ports for the USART as
UART0 -> USART0, UART1 -> USART1, SPI0 -> USART0, etc.   The x2 and x5
processors have many more ports than the simple USART implementation.
Further, it is a platform specific prerogative to determine what ports
are called what and how they are wired.

The original x2 USCI port followed the precedent established by the x1
USART naming.  But after examining the x5 implementation it became clear
that was a mistake.  Rather a simple naming that ties the function to
the actual h/w is superior.