msp430: clean up arbitrated interface at the usci level.
Motivation: The naming convention was unclear and applied inconsistently with
some of the new x2 processor code.  Clean it up.  Much of this was a hold over
from the old x1 USART code that provided a model for the original x2 code.

Originally, Msp430Usci{A,B}{0,1}C and Msp430UsciShare{A,B}{0,1}P was a pair that
provided an arbitrated interface to the low level Hpl Usci interfaces.
Make the names consistent.

Msp430Usci{A,B}{0,1}C -> Msp430UsciArb{A,B}{0,1}C
Msp430UsciShare{A,B}{0,1}P -> Msp430UsciArb{A,B}{0,1}P

Also change Msp430UsciShareP -> Msp430UsciArbIntP.    This module is simply
an interrupt steering linkage.   The new name is significantly better.