msp430 dma: nuke ABORT.
Motivation: ABORT gets set when DMA has been nuked by a NMI occuring.  This can only happen
if ENNMI is set in the opControl word.  ABORT impacted the DMA interfaces by causing there
to be an error return to the transferDone signal.   Note if one is really doing something
off the NMI then shutting down the DMA to focus the CPU on handling the NMI is a reasonable
thing to do.

But there are no known uses of NMI.  Further this complicates the interfaces to the DMA without
a sufficiently good payback.  And no known handles for the error return exist.  In other words
interface was added without there really being a good justification for the complexity.

Now that said, it is a good idea to check to make sure that the DMA has actually completed and
hasn't failed in some fashion.   DMA can hang or fail.  ABORT is an example of the DMA failing
to complete.   So the ABORT error interface above handles part of the problem.

A more complete solution requires any DMA transaction to be wrapped by a timer to detect
timeouts.  This same mechanism would also detect non-completion due to an ABORT.  In other
words the ABORT error interface doesn't really buy anything and complicates the interface
for no good reason.

See ya ABORT.