msp430 dma: nuke error return from setTrigger
Motivation:  setTrigger could only fail if DMAEN was set.  And no calls to setTrigger
actually checked the return value.  Now if DMAEN is set, changing the trigger is ignored
by the h/w.   But if someone is using setTrigger they should a) know what they are doing
(and so will know that DMAEN had better be off) and b) should check their work, ie. actually
look to see that the DMA configuration they are setting up is done correctly.  Now once
it is done correctly it won't ever fail so why have the extra complexity.

In other words there really isn't a good reason to generate an error return.