msp430: telos platforms were overriding optimization, use default.
According to rumor there was a bug in the 3.2.3 or some other toolchain in
use at the time that required -O after the -Os to override it.

However, some applications (say PppRouter) need space optimization inorder
to fit, so specifing -Os -O is wrong.

This fix deletes the override and just uses the default of -Os.