clean up of ieee154_header_t caused more problems then it was worth.  Revert.
An attempt was made to pull all the salient bits of the ieee 802.15.4 packet
headers into tos/types/Ieee154.h but it failed.   The bigest problem was a
strange compiler problem that caused typedef nx_struct ieee154_simple_header_t
to barf.  Unclear why.

Also Miklos is doing a major piece of work to allow support for 16 and 64 bit
address via accessors and packers.   This will change all of this anyway so it
isn't worth sorting out.