USCI Pin mapping occurs in platforms/<platform>/hardware/usci/PlatformUsciMapC.nc
PlatformUsciMapC defines the mapping between USCI pins and where the pins
actually are on the chip.   This is done on a per platform basis because
a CPU is herently a platform thing and also we only want to map what pins
we are actually using....

Clean up various places where pin mapping erroneously got stuck in the
msp430 core usci files.   The core files are pin mapping agnostic and
putting pin mappings out there actually cause problems.

Change the mapping file from PlatformUsciInitC to PlatformUsciMapC.
This better reflects what the purpose of this files is.  It doesn't
actually initilize anything.